requirements.txt
requests>=2.31.0
mcp>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
opensearch-py>=2.4.0